Update readme to include a list of error messages
After running this once I woke up, I realized that some of the error messages could be confusing, in particular receiving a left bracket when you do not expect to can actually mean a poorly formatted operation name.  This should hopefully provide a bit more clarity for the time being.